FIX: 의존성 사이클 문제 해결
