+++ src/app/mypage/setting/alarm.tsx
@@ -1,5 +1,5 @@
-import { SettingsView } from './page'
 import { TopBar, Switch } from '@eolluga/eolluga-ui'
+import { SettingsView } from './page'
 
 export default function AlarmPage({ handleChangeView }: { handleChangeView: (value: SettingsView) => void }) {
   return (
