+++ src/app/mypage/setting/page.tsx
@@ -1,12 +1,12 @@
 'use client'
 
-import { TopBar, Icon } from '@eolluga/eolluga-ui'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
-import AlarmPage from './alarm'
-import PrivacyPage from './privacy'
+import { TopBar, Icon } from '@eolluga/eolluga-ui'
 import privacyPolicyData from '@/data/privacyPolicy.json'
+import AlarmPage from './alarm'
 import AppInfoPage from './appInfo'
+import PrivacyPage from './privacy'
 import BusinessInfoPage from './businessInfo'
 import LogOutModal from './logOut'
 import QuitPage from './quit'
